requirements.txt
Django               # Web framework
django-crispy-forms # Form styling
django-environ       # Environment variables management
djangorestframework # RESTful API framework
Pillow               # Image processing library
django-allauth       # Authentication and account management
django-storages      # Storage backends for file uploads (e.g., Amazon S3)
boto3                # AWS SDK for Python
django-countries     # Provides country choices for forms
django-ckeditor      # Rich text editor for content management
django-debug-toolbar # Debugging tool for Django applications
requests
paypalrestsdk
django-paypal